refact(Transactions): reorganize files
