+++ src/model-types.ts
@@ -1,8 +1,3 @@
-import { Money } from 'utils/money';
-import { BankAccount } from 'services/bank-accounts/types';
-import { Customer } from 'services/customers/types';
-import { Category } from 'services/categories/types';
-
 export enum Status {
   Invalid = 'INVALID',
   Pending = 'PENDING',
@@ -15,35 +10,4 @@ export interface Pagination {
   previous?: number;
   pages: number;
   next?: number;
-}
-
-export interface TransferOut {
-  id:          number;
-  amount:      Money;
-  category:    Category;
-  bankAccount: BankAccount;
-  date?:       Date;
-  comment?:    string;
-}
-
-export interface Transaction {
-  id:          number;
-  amount:      Money;
-  category:    Category;
-  bankAccount: BankAccount;
-  customer?:   Customer;
-  date?:       Date;
-  comment?:    string;
-  isViewed:    boolean;
-
-  transfer_out?: TransferOut;
-}
-
-export interface TransactionParams {
-  amount?:        string;
-  categoryId?:    number;
-  customerId?:    number;
-  bankAccountId?: number;
-  date?:          Date;
-  comment?:       string;
 }
