+++ src/sagas/index.ts
@@ -4,7 +4,7 @@ import authSaga          from 'services/auth/saga.js';
 import balancesSaga      from './balances';
 import bankAccountsSaga  from 'services/bank-accounts/saga.js';
 import organizationsSaga from 'services/organizations/saga.js';
-import transactionsSaga  from './transactions';
+import transactionsSaga  from 'services/transactions/saga.js';
 import customersSaga     from 'services/customers/saga.js';
 import categoriesSaga    from 'services/categories/saga.js';
 import membersSaga       from './members';
